+++ sdk/confidentialledger/azure-confidentialledger/azure/confidentialledger/receiptverification/__init__.py
@@ -4,14 +4,13 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------
 
-from .receipt_verification import verify_receipt, verify_receipt_from_dict
-from .models import Receipt, LeafComponents, ProofElement
+from ._receipt_verification import verify_receipt
+from ._models import Receipt, LeafComponents, ProofElement
 from ._version import VERSION
 
 __version__ = VERSION
 __all__ = [
     "verify_receipt",
-    "verify_receipt_from_dict",
     "Receipt",
     "LeafComponents",
     "ProofElement",