+++ sdk/confidentialledger/azure-confidentialledger/tests/receiptverification/test_models.py
@@ -7,7 +7,7 @@
 
 import pytest
 
-from azure.confidentialledger.receiptverification.models import (
+from azure.confidentialledger.receiptverification._models import (
     LeafComponents,
     ProofElement,
     Receipt,